ai-engine/requirements.txt
# AI Engine Requirements
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-json-logger==2.0.7
python-dotenv==1.0.0
httpx==0.26.0

# Vector store and embeddings
faiss-cpu==1.7.4
numpy==1.24.3

# LLM clients
# Ollama via httpx (already included)
# openai==1.10.0  # Optional
